Re-factor test functions into base class
