+++ cllc-public-app/Test/LoginControllerTests.cs
@@ -16,75 +16,29 @@ using Gov.Lclb.Cllb.Public.Models;
 
 namespace Gov.Lclb.Cllb.Public.Test
 {
-    public class LoginControllerTests : ApiIntegrationTestBaseAnonymous
+    public class LoginControllerTests : ApiIntegrationTestBaseWithLogin
     {
         [Fact]
         public async System.Threading.Tasks.Task DefaultUserIsAnonymous()
         {
-            var request = new HttpRequestMessage(HttpMethod.Get, "/api/user/current");
-            var response = await _client.SendAsync(request);
-			Assert.Equal(response.StatusCode, HttpStatusCode.Unauthorized);
-			string _discard = await response.Content.ReadAsStringAsync();
-        }
-
-		[Fact]
-        public async System.Threading.Tasks.Task LoginAsDefaultUserSetsCurrentUser()
-        {
-            var request = new HttpRequestMessage(HttpMethod.Get, "/login/token/default");
-            var response = await _client.SendAsync(request);
-			Assert.Equal(response.StatusCode, HttpStatusCode.Found);
-			string _discard = await response.Content.ReadAsStringAsync();
-
-			_client.DefaultRequestHeaders.Add("DEV-USER", "TMcTesterson");
-
-			var request2 = new HttpRequestMessage(HttpMethod.Get, "/api/user/current");
-            var response2 = await _client.SendAsync(request2);
-            response2.EnsureSuccessStatusCode();
-            string jsonString = await response2.Content.ReadAsStringAsync();
-            // account and contact id's will be different ...
-			//Assert.Equal(jsonString, "{\n  \"id\": \"TMcTesterson\",\n  \"name\": \"TMcTesterson TestUser\",\n  \"firstname\": \"TMcTesterson\",\n  \"lastname\": \"TestUser\",\n  \"email\": null,\n  \"businessname\": \"TMcTesterson TestBusiness\",\n  \"isNewUser\": true,\n  \"isContactCreated\": false,\n  \"isAccountCreated\": false,\n  \"isBceidConfirmed\": false,\n  \"contactid\": \"e901654e-34e6-4e16-9a5f-980bc08bbee7\",\n  \"accountid\": \"3ebb1d3a-0dfb-4fd2-b6f2-10cc08d95283\"\n}");
-
-            ViewModels.User user = JsonConvert.DeserializeObject<ViewModels.User>(jsonString);
-
-            // Verify the Default development user.
-			Assert.Equal(user.name, "TMcTesterson TestUser");
-			Assert.True(user.isNewUser);
+			await GetCurrentUserIsUnauthorized();
         }
 
         [Fact]
-        public async System.Threading.Tasks.Task LoginThenLogoutIsAnonymous()
+        public async System.Threading.Tasks.Task LoginSetsCurrentUserThenLogoutIsAnonymous()
         {
-            var request = new HttpRequestMessage(HttpMethod.Get, "/login/token/default");
-            var response = await _client.SendAsync(request);
-            Assert.Equal(response.StatusCode, HttpStatusCode.Found);
-            string _discard = await response.Content.ReadAsStringAsync();
+			await LoginAsDefault();
 
-            _client.DefaultRequestHeaders.Add("DEV-USER", "TMcTesterson");
-
-            var request2 = new HttpRequestMessage(HttpMethod.Get, "/api/user/current");
-            var response2 = await _client.SendAsync(request2);
-            response2.EnsureSuccessStatusCode();
-            string jsonString = await response2.Content.ReadAsStringAsync();
-            // account and contact id's will be different ...
-            //Assert.Equal(jsonString, "{\n  \"id\": \"TMcTesterson\",\n  \"name\": \"TMcTesterson TestUser\",\n  \"firstname\": \"TMcTesterson\",\n  \"lastname\": \"TestUser\",\n  \"email\": null,\n  \"businessname\": \"TMcTesterson TestBusiness\",\n  \"isNewUser\": true,\n  \"isContactCreated\": false,\n  \"isAccountCreated\": false,\n  \"isBceidConfirmed\": false,\n  \"contactid\": \"e901654e-34e6-4e16-9a5f-980bc08bbee7\",\n  \"accountid\": \"3ebb1d3a-0dfb-4fd2-b6f2-10cc08d95283\"\n}");
-
-            ViewModels.User user = JsonConvert.DeserializeObject<ViewModels.User>(jsonString);
-
+			string jsonString = await GetCurrentUser();
+            
             // Verify the Default development user.
+			ViewModels.User user = JsonConvert.DeserializeObject<ViewModels.User>(jsonString);
             Assert.Equal(user.name, "TMcTesterson TestUser");
             Assert.True(user.isNewUser);
 
-			var request3 = new HttpRequestMessage(HttpMethod.Get, "/login/cleartoken");
-            var response3 = await _client.SendAsync(request3);
-            Assert.Equal(response3.StatusCode, HttpStatusCode.Found);
-            string _discard3 = await response3.Content.ReadAsStringAsync();
+			await Logout();
 
-			_client.DefaultRequestHeaders.Remove("DEV-USER");
-
-			var request4 = new HttpRequestMessage(HttpMethod.Get, "/api/user/current");
-            var response4 = await _client.SendAsync(request4);
-            Assert.Equal(response4.StatusCode, HttpStatusCode.Unauthorized);
-            string _discard4 = await response4.Content.ReadAsStringAsync();
+			await GetCurrentUserIsUnauthorized();
         }
 	}
 }
